+++ src/core/resources/assets.rs
@@ -3,8 +3,14 @@ use bevy_asset_loader::prelude::*;
 
 #[derive(AssetCollection, Resource)]
 pub struct SpriteAssets {
-    #[asset(path = "space/ships/playerShip2_blue.png")]
+    #[asset(path = "space/ships/player.png")]
     pub player_ship: Handle<Image>,
+
+    #[asset(path = "space/ships/enemy.png")]
+    pub enemy_ship: Handle<Image>,
+
+    #[asset(path = "space/ships/bullet.png")]
+    pub bullet: Handle<Image>,
 
     #[asset(texture_atlas(tile_size_x = 500., tile_size_y = 500., columns = 25, rows = 5))]
     #[asset(path = "space/celestials/star-pixelplanet.png")]
@@ -27,11 +33,14 @@ pub struct SpriteAssets {
 
 #[derive(AssetCollection, Resource)]
 pub struct AudioAssets {
-    #[asset(path = "sound/Beat Mekanik - Lightspeed.ogg")]
+    #[asset(path = "sound/music/Beat Mekanik - Lightspeed.ogg")]
     pub title_music: Handle<AudioSource>,
 
-    #[asset(path = "sound/Kirk Osamayo - Space Dust.ogg")]
+    #[asset(path = "sound/music/Kirk Osamayo - Space Dust.ogg")]
     pub ambience: Handle<AudioSource>,
+
+    #[asset(path = "sound/sfx/impactPlank_medium_000.ogg")]
+    pub gun: Handle<AudioSource>,
 }
 
 #[derive(AssetCollection, Resource)]