+++ src/ui/menus/credits.rs
@@ -60,7 +60,7 @@ fn setup(mut commands: Commands, ui: Res<UiAssets>) {
             },));
             parent.spawn((TextBundle {
                 style: Style {
-                    margin: UiRect::top(Val::Px(50.)),
+                    margin: UiRect::top(Val::Px(25.)),
                     ..default()
                 },
                 text: Text::from_section(
@@ -101,7 +101,7 @@ fn setup(mut commands: Commands, ui: Res<UiAssets>) {
             },));
             parent.spawn((TextBundle {
                 style: Style {
-                    margin: UiRect::top(Val::Px(50.)),
+                    margin: UiRect::top(Val::Px(25.)),
                     ..default()
                 },
                 text: Text::from_section(
@@ -132,6 +132,32 @@ fn setup(mut commands: Commands, ui: Res<UiAssets>) {
             parent.spawn((TextBundle {
                 text: Text::from_section(
                     "kenney.nl, CC0",
+                    TextStyle {
+                        font: ui.font.clone(),
+                        font_size: 14.0,
+                        color: Color::rgb_u8(0xAA, 0xAA, 0xAA),
+                    },
+                ),
+                ..default()
+            },));
+            parent.spawn((TextBundle {
+                style: Style {
+                    margin: UiRect::top(Val::Px(25.)),
+                    ..default()
+                },
+                text: Text::from_section(
+                    "Ship Mixer by Kenney",
+                    TextStyle {
+                        font: ui.font.clone(),
+                        font_size: 20.0,
+                        color: Color::rgb_u8(0xCC, 0xCC, 0xCC),
+                    },
+                ),
+                ..default()
+            },));
+            parent.spawn((TextBundle {
+                text: Text::from_section(
+                    "kenney.nl",
                     TextStyle {
                         font: ui.font.clone(),
                         font_size: 14.0,
@@ -194,7 +220,7 @@ fn setup(mut commands: Commands, ui: Res<UiAssets>) {
             },));
             parent.spawn((TextBundle {
                 style: Style {
-                    margin: UiRect::top(Val::Px(50.)),
+                    margin: UiRect::top(Val::Px(25.)),
                     ..default()
                 },
                 text: Text::from_section(
@@ -259,5 +285,46 @@ fn setup(mut commands: Commands, ui: Res<UiAssets>) {
                 ),
                 ..default()
             },));
+            // parent.spawn((TextBundle {
+            //     style: Style {
+            //         margin: UiRect::top(Val::Px(25.)),
+            //         ..default()
+            //     },
+            //     text: Text::from_section(
+            //         "Audio".to_ascii_uppercase(),
+            //         TextStyle {
+            //             font: ui.font.clone(),
+            //             font_size: 24.0,
+            //             color: Color::rgb_u8(0x00, 0x88, 0x88),
+            //         },
+            //     ),
+            //     ..default()
+            // },));
+            // parent.spawn((TextBundle {
+            //     style: Style {
+            //         margin: UiRect::top(Val::Px(25.)),
+            //         ..default()
+            //     },
+            //     text: Text::from_section(
+            //         "Impact Sounds by Kenney",
+            //         TextStyle {
+            //             font: ui.font.clone(),
+            //             font_size: 20.0,
+            //             color: Color::rgb_u8(0xCC, 0xCC, 0xCC),
+            //         },
+            //     ),
+            //     ..default()
+            // },));
+            // parent.spawn((TextBundle {
+            //     text: Text::from_section(
+            //         "kenney.nl, CC0",
+            //         TextStyle {
+            //             font: ui.font.clone(),
+            //             font_size: 14.0,
+            //             color: Color::rgb_u8(0xAA, 0xAA, 0xAA),
+            //         },
+            //     ),
+            //     ..default()
+            // },));
         });
 }